requirements.txt
pytest~=7.4.3
numpy~=1.24.4
matplotlib~=3.7.3
Pillow~=10.1.0